\name{scoop-internal}

\alias{bfgs.lm.coo}
\alias{bfgs.lm.grp}
\alias{bfgs.lm.las}
\alias{bfgs.lrm.coo}
\alias{bfgs.lrm.grp}
\alias{bfgs.lrm.las}
\alias{df.scoop}
\alias{dL.coop.binomial}
\alias{dL.coop.gaussian}
\alias{dL.group.binomial}
\alias{dL.group.gaussian}
\alias{dL.lasso.binomial}
\alias{dL.lasso.gaussian}
\alias{err.deviance}
\alias{fista.lm.coo}
\alias{fista.lm.grp}
\alias{fista.lm.las}
\alias{is.scoop}
\alias{is.cv.scoop}
\alias{is.intercept}
\alias{ista.lm.coo}
\alias{ista.lm.grp}
\alias{ista.lm.las}
\alias{L.coop.binomial}
\alias{L.coop.gaussian}
\alias{L.group.binomial}
\alias{L.group.gaussian}
\alias{L.lasso.binomial}
\alias{L.lasso.gaussian}
\alias{lm.coo}
\alias{lm.grp}
\alias{lm.las}
\alias{lrm.coo}
\alias{lrm.grp}
\alias{lrm.las}
\alias{group.norm}
\alias{group.norm.rep}
\alias{group.norm.unordered}
\alias{reg.fitting}
\alias{sigmoid}

\title{'scoop' internal functions}
\description{
  Internal functions.
}

\details{
  internal functions: not to be called by the user.
}

\author{Julien Chiquet}

\keyword{internal}