% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCOOP-class.R
\name{SCOOP}
\alias{SCOOP}
\title{R6 abstract class for a generic sparse modelmodel}
\description{
R6 abstract class for a generic sparse modelmodel

R6 abstract class for a generic sparse modelmodel
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{group}}{vector of penalty grouping, when applicable}

\item{\code{family}}{type of loss (aka family)}

\item{\code{penalty}}{type of penalty function}

\item{\code{penalty_level}}{vector of amounts of penalty applied (aka lambda)}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{d}}{number of variables}

\item{\code{k}}{number of groups}

\item{\code{group}}{vector of penalty grouping, when applicable}

\item{\code{family}}{type of loss (aka family)}

\item{\code{penalty}}{type of penalty function}

\item{\code{penalty_level}}{vector of amounts of penalty applied (aka lambda)}

\item{\code{penalty_weights}}{vector of weights applied at the group level (or coefficients level if no grouping)}

\item{\code{fitted}}{a matrix: fitted values}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SCOOP-new}{\code{SCOOP$new()}}
\item \href{#method-SCOOP-plot}{\code{SCOOP$plot()}}
\item \href{#method-SCOOP-predict}{\code{SCOOP$predict()}}
\item \href{#method-SCOOP-show}{\code{SCOOP$show()}}
\item \href{#method-SCOOP-clone}{\code{SCOOP$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SCOOP-new"></a>}}
\if{latex}{\out{\hypertarget{method-SCOOP-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCOOP$new(data, family)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SCOOP-plot"></a>}}
\if{latex}{\out{\hypertarget{method-SCOOP-plot}{}}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCOOP$plot(
  x,
  y,
  xvar = "lambda",
  yvar = "coefficients",
  main = paste(self$penalty_type, " path (", yvar, ")", sep = ""),
  crit = NULL,
  log_scale = TRUE,
  labels = NULL,
  col = c(1 + switch(yvar, group = 1:self$k, self$grouping)),
  lty = switch(yvar, group = self$k, self$grouping),
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SCOOP-predict"></a>}}
\if{latex}{\out{\hypertarget{method-SCOOP-predict}{}}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCOOP$predict(newx = self$data$x, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SCOOP-show"></a>}}
\if{latex}{\out{\hypertarget{method-SCOOP-show}{}}}
\subsection{Method \code{show()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCOOP$show()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SCOOP-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SCOOP-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCOOP$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
